backend/requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.1.7
django-bootstrap4==2.3.1
djangorestframework==3.12.2
pytz==2021.1
soupsieve==2.2
sqlparse==0.4.1
django-environ==0.4.5
pymysql==1.0.2
cryptography==3.4.6